xwayland: unredirect sub-windows when root destroyed
This ensures that we unredirect root sub-windows when the root window is
being destroyed. We need to do this explicitly rather than rely on the
the CompositeClientSubwindows resource being freed automatically because
that may not happen until after the root window has already been freed
resulting in uninitialized memory access.