exa: Perform texture conversions before falling back
This is highly inelegant, and nothing more than a hack to prevent
fallbacks due to unhandled SourcePicts. It is a mixed blessing, but it
maybe useful to mitigate some extreme cases where triggering a fallback
is a performance nightmare.

On a SNB i5-2500 with a Radeon HD5770:

Speedups
========
 xlib                  evolution  3369.15 (0.25%) -> 2274.31 (0.20%):  1.48x speedup
 xlib         firefox-chalkboard  53283.75 (0.30%) -> 42165.15 (3.18%):  1.26x speedup
 xlib       firefox-planet-gnome  9933.11 (0.63%) -> 8435.16 (0.66%):  1.18x speedup
 xlib           firefox-fishtank  21460.93 (0.06%) -> 19448.76 (0.10%):  1.10x speedup
 xlib          firefox-talos-gfx  16931.70 (0.17%) -> 15672.78 (0.23%):  1.08x speedup
 xlib             swfdec-youtube  2046.82 (0.01%) -> 1919.43 (0.10%):  1.07x speedup

Slowdowns
=========
 xlib          firefox-talos-svg  25700.24 (0.03%) -> 29653.74 (0.04%):  1.15x slowdown
 xlib              midori-zoomed  1461.05 (0.17%) -> 1916.89 (2.36%):  1.31x slowdown

As this is not the order of magnitude change one would expect from avoid
a fallback, I consider this patch to be irrelevant for the time being.